refactor: entity 패키지를 삭제하고 도메인으로 이동
